+++ packages/react-components/react-button/stories/MenuButton/MenuButtonSizeSmall.stories.tsx
@@ -1,0 +1,74 @@
+import * as React from 'react';
+import { bundleIcon, CalendarMonthFilled, CalendarMonthRegular } from '@fluentui/react-icons';
+import {
+  makeStyles,
+  Menu,
+  MenuButton,
+  MenuItem,
+  MenuList,
+  MenuPopover,
+  MenuTrigger,
+  Tooltip,
+} from '@fluentui/react-components';
+
+const CalendarMonth = bundleIcon(CalendarMonthFilled, CalendarMonthRegular);
+
+const useStyles = makeStyles({
+  wrapper: {
+    columnGap: '15px',
+    display: 'flex',
+  },
+});
+
+export const SizeSmall = () => {
+  const styles = useStyles();
+
+  return (
+    <div className={styles.wrapper}>
+      <Menu>
+        <MenuTrigger>
+          <MenuButton size="small">Small</MenuButton>
+        </MenuTrigger>
+
+        <MenuPopover>
+          <MenuList>
+            <MenuItem>Item a</MenuItem>
+            <MenuItem>Item b</MenuItem>
+          </MenuList>
+        </MenuPopover>
+      </Menu>
+
+      <Menu>
+        <MenuTrigger>
+          <MenuButton icon={<CalendarMonth />} size="small">
+            Small with calendar icon
+          </MenuButton>
+        </MenuTrigger>
+
+        <MenuPopover>
+          <MenuList>
+            <MenuItem>Item a</MenuItem>
+            <MenuItem>Item b</MenuItem>
+          </MenuList>
+        </MenuPopover>
+      </Menu>
+
+      <Menu>
+        <MenuTrigger>
+          <Tooltip content="Small with calendar icon only" relationship="label">
+            <MenuButton icon={<CalendarMonth />} size="small" />
+          </Tooltip>
+        </MenuTrigger>
+
+        <MenuPopover>
+          <MenuList>
+            <MenuItem>Item a</MenuItem>
+            <MenuItem>Item b</MenuItem>
+          </MenuList>
+        </MenuPopover>
+      </Menu>
+    </div>
+  );
+};
+
+SizeSmall.storyName = 'Size: small';