+++ packages/react-components/react-progress/src/components/Progress/Progress.types.ts
@@ -2,30 +2,13 @@ import type { ComponentProps, ComponentState, Slot } from '@fluentui/react-utili
 
 export type ProgressSlots = {
   /**
-   * The root of the Progress
-   * The root slot receives the `className` and `style` specified directly on the `<Progress>`.
+   * The track behind the progress bar
    */
   root: NonNullable<Slot<'div'>>;
   /**
-   * The title of the Progress.
-   * The label slot receives the styling related to the title associated with the Progress.
-   */
-  label?: Slot<'span'>;
-  /**
-   * The animated slot of the Progress
-   * The bar slot receives the styling related to the loading bar associated with the Progress
+   * The filled portion of the progress bar. Animated in the indeterminate state, when no value is provided.
    */
   bar?: NonNullable<Slot<'div'>>;
-  /**
-   * The track slot of the Progress
-   * The track slot receives the styling related to the loading bar track associated with the Progress
-   */
-  track?: NonNullable<Slot<'div'>>;
-  /**
-   * The description slot of the Progress
-   * The description slot receives the styling related to the description associated with the Progress
-   */
-  description?: Slot<'span'>;
 };
 
 /**
@@ -33,23 +16,38 @@ export type ProgressSlots = {
  */
 export type ProgressProps = Omit<ComponentProps<ProgressSlots>, 'size'> & {
   /**
-   * Prop to set whether the Progress is determinate or indeterminate
-   * @default false
+   * The shape of the bar and track.
+   * @default 'rounded'
    */
-  indeterminate?: boolean;
+  shape?: 'rounded' | 'rectangular';
   /**
-   * Percentage of the operation's completeness, numerically between 0 and 100.
+   * A decimal number between `0` and `1` (or between `0` and `max` if given),
+   * which specifies how much of the task has been completed.
+   *
+   * If `undefined` (default), the Progress will display an **indeterminate** state.
    */
-  percentComplete?: number;
+  value?: number;
+  /**
+   * The maximum value, which indicates the task is complete.
+   * The progress bar will be full when `value` equals `max`.
+   * @default 1
+   */
+  max?: number;
   /**
    * The thickness of the Progress bar
    * @default 'medium'
    */
   thickness?: 'medium' | 'large';
+
+  /**
+   * The status of the progress bar. Changes the color of the bar.
+   */
+  validationState?: 'success' | 'warning' | 'error';
 };
 
 /**
  * State used in rendering Progress
  */
 export type ProgressState = ComponentState<ProgressSlots> &
-  Required<Pick<ProgressProps, 'indeterminate' | 'percentComplete' | 'thickness'>>;
+  Required<Pick<ProgressProps, 'max' | 'shape' | 'thickness'>> &
+  Pick<ProgressProps, 'value' | 'validationState'>;