+++ packages/react-components/react-progress/src/stories/Progress/index.stories.tsx
@@ -4,9 +4,11 @@ import descriptionMd from './ProgressDescription.md';
 import bestPracticesMd from './ProgressBestPractices.md';
 
 export { Default } from './ProgressDefault.stories';
-export { Appearance } from './ProgressAppearance.stories';
+export { Shape } from './ProgressShape.stories';
 export { Thickness } from './ProgressBarThickness.stories';
 export { Indeterminate } from './ProgressIndeterminate.stories';
+export { ValidationState } from './ProgressValidationState.stories';
+export { Max } from './ProgressMax.stories';
 
 export default {
   title: 'Preview Components/Progress',