+++ packages/react-components/react-table/src/stories/Table/index.stories.tsx
@@ -13,9 +13,8 @@ export { MultipleSelect } from './MultipleSelect.stories';
 export { SingleSelect } from './SingleSelect.stories';
 export { MultipleSelectControlled } from './MultipleSelectControlled.stories';
 export { SingleSelectControlled } from './SingleSelectControlled.stories';
-export { CellNavigationMode } from './CellNavigationMode.stories';
-export { RowNavigationMode } from './RowNavigationMode.stories';
-export { CompositeNavigationMode } from './CompositeNavigationMode.stories';
+export { CellNavigation } from './CellNavigation.stories';
+export { SubtleSelection } from './SubtleSelection.stories';
 
 export default {
   title: 'Preview Components/Table',