+++ packages/react-components/react-toolbar/src/components/ToolbarToggleButton/ToolbarToggleButton.test.tsx
@@ -14,7 +14,11 @@ describe('ToolbarToggleButton', () => {
   // TODO add more tests here, and create visual regression tests in /apps/vr-tests
 
   it('renders a default state', () => {
-    const result = render(<ToolbarToggleButton>Default ToolbarToggleButton</ToolbarToggleButton>);
+    const result = render(
+      <ToolbarToggleButton name="name" value="value">
+        Default ToolbarToggleButton
+      </ToolbarToggleButton>,
+    );
     expect(result.container).toMatchSnapshot();
   });
 });