+++ packages/react-examples/src/react/Callout/Callout.Basic.Example.tsx
@@ -28,7 +28,7 @@ export const CalloutBasicExample: React.FunctionComponent = () => {
           onDismiss={toggleIsCalloutVisible}
           setInitialFocus
         >
-          <Text block variant="xLarge" className={styles.title} id={labelId}>
+          <Text as="h1" block variant="xLarge" className={styles.title} id={labelId}>
             Callout title here
           </Text>
           <Text block variant="small" id={descriptionId}>